// apps/web-emoney/src/graphql/types/general_types.tsx
export type ParamsQuery = { stakingParams: Array<{ __typename?: 'staking_params', params: any }>, slashingParams: Array<{ __typename?: 'slashing_params', params: any }>, mintParams: Array<{ __typename?: 'mint_params', params: any }>, distributionParams: Array<{ __typename?: 'distribution_params', params: any }>, govParams: Array<{ __typename?: 'gov_params', depositParams: any, tallyParams: any, votingParams: any }>, inflationRateParams: Array<{ __typename?: 'emoney_inflation', inflation: any }>, gasPriceParams: Array<{ __typename?: 'emoney_gas_prices', gas_prices: any }> };